refactor: move model catalog and HTTP layer to leaf modules (move-only)
- src/model_catalog.ts: MODEL_VARIANTS + findVariant (pure data)
- src/api_client.ts: fetchWithRetry, formatApiError, BASE_URL,
  BALANCE_URL, BalanceInfo (vscode-free — the retry wrapper becomes
  unit-testable with a monkey-patched global fetch, which the next
  commit uses)

All three moved blocks verified byte-identical to the pre-move
originals (automated comparison against git HEAD). Deliberately NOT
moved, per the adversarial plan review: refreshBalance (its HTTP call
interleaves with session-anchor mutation, silent-mode log routing, and
status-bar refresh — extracting "the HTTP part" would be new code, not
a move), notifyApiError and showContextOverflowGuidance (vscode.window
throughout), and the picker-info assembly (ThemeIcon + secrets gating).
The unit_tool_limit text pin on out/provider.js is unaffected — the
guard call site stays in provider.ts.
